Remove TrilinosWrappers::VectorBase
